api/requirements.txt
flask-restplus
Flask-SQLAlchemy
setuptools
